+++ src.wsjcpp/wsjcpp_arguments/wsjcpp_arguments.h
@@ -67,12 +67,13 @@ class WsjcppArgumentProcessor {
         WsjcppArgumentParameter *findRegisteredParameterArgument(const std::string &sArgumentName);
 
         bool hasRegisteredArgumentName(const std::string &sArgumentName);
-        
+
         bool getValueOfParam(const std::string &sArgumentName);
         int help(
             const std::vector<std::string> &vRoutes, 
             const std::vector<std::string> &vSubParams
         );
+        bool hasMoreOptions();
 
         virtual bool applySingleArgument(const std::string &sProgramName, const std::string &sArgumentName);
         virtual bool applyParameterArgument(const std::string &sProgramName, const std::string &sArgumentName, const std::string &sValue);
@@ -112,6 +113,7 @@ class WsjcppArguments {
     public:
         WsjcppArguments(int argc, const char* argv[], WsjcppArgumentProcessor *pRoot);
         ~WsjcppArguments();
+        void enablePrintAutoHelp(bool bEnablePrintAutoHelp);
         int exec();
 
     private:
@@ -131,7 +133,9 @@ class WsjcppArguments {
         std::string TAG;
         std::vector<std::string> m_vArguments;
         std::string m_sProgramName;
+        bool m_bEnablePrintAutoHelp;
         std::vector<WsjcppArgumentProcessor *> m_vProcessors;
+        
 };
 
 // ---------------------------------------------------------------------