+++ src/service_locator/employ_config.cpp
@@ -13,7 +13,6 @@
 #include <wsjcpp_yaml.h>
 #include <wsjcpp_validators.h>
 #include <employ_team_logos.h>
-#include <wsjcpp_resources_manager.h>
 #include <sys/stat.h>
 #include <stdio.h>
 
@@ -131,7 +130,7 @@ void Ctf01dTeamDef::setId(const std::string &sTeamId){
 
 // ---------------------------------------------------------------------
 
-std::string Ctf01dTeamDef::id() const {
+std::string Ctf01dTeamDef::getId() const {
     return m_sTeamID;
 }
 
@@ -143,7 +142,7 @@ void Ctf01dTeamDef::setName(const std::string &sName){
 
 // ---------------------------------------------------------------------
 
-std::string Ctf01dTeamDef::name() const {
+std::string Ctf01dTeamDef::getName() const {
     return m_sName;
 }
 
@@ -273,6 +272,9 @@ bool EmployConfig::deinit() {
 // ---------------------------------------------------------------------
 
 void EmployConfig::setWorkDir(const std::string &sWorkDir) {
+    if (m_sWorkDir != "" && m_sWorkDir != sWorkDir) {
+        std::cout << "Changed work-dir to '" + sWorkDir + "'" << std::endl;
+    }
     m_sWorkDir = sWorkDir;
     m_sScoreboardHtmlFolder = m_sWorkDir + "/html"; // default value
 }
@@ -852,7 +854,7 @@ bool EmployConfig::readTeamsConf(WsjcppYaml &yamlConfig) {
         }
 
         for (unsigned int i = 0; i < m_vTeamsConf.size(); i++) {
-            if (m_vTeamsConf[i].id() == sTeamId) {
+            if (m_vTeamsConf[i].getId() == sTeamId) {
                 WsjcppLog::err(TAG, "Already registered team with id " + sTeamId);
                 return false;
             }